Update README.md
I made 64bit audio. The reason why they haven't come out with it yet is because 32bit audio has to have a sample rate of 16 to 24bit to be compatible with HDMI and Bluetooth so that is why they haven't come out with it yet. Now the jack on your phone doesn't have a sample rate it's analog and your Dad taught me that analog is the best sound. So what I compiled is 64bit audio with 64bit math tables and 32bit sample rate. It will only work for auxiliary output. What it actually is it is a ALSA plugin the kernel set's the permissions for music file playback with the ALSA library. Now what I built takes the data decoded from a music file and encodes it as DTS Digital theater sound. The music file is not affected. The file format doesn't matter .flac .mp3 .m4a they all pretty much sound the same Killer. I got the source code and the ALSA headers for me to build it. It was a miracle. I have a repository on GitHub with the 32bit versions. I have no plans to release these versions. I'm trying to get in touch with the developer to go further and build a release versions of the app for purchase. There is nothing like it. It won't work for streaming apps like YouTube and YouTube music and tidal. But it will work for Spotify and Iphone music purchases. The best way is Android with a 128GB SD card and Poweramp music player and the retro metallic skin both are paid under $10 bucks for both it has the best user interface. Please forward to your dad he will be very interested.



Best regards
Ryan Johnson